bot/requirements.txt
#havent tested this yet, most of this requirements are not necessary
aiohttp            3.9.1
aiosignal          1.3.1
annotated-types    0.6.0
anyio              4.2.0
attrs              23.2.0
blinker            1.8.2
certifi            2023.11.17
charset-normalizer 3.3.2
click              8.1.7
colorama           0.4.6
discord            2.3.2
discord.py         2.3.2
distro             1.9.0
Flask              3.0.3
frozenlist         1.4.1
h11                0.14.0
httpcore           1.0.2
httpx              0.26.0
idna               3.6
itsdangerous       2.2.0
Jinja2             3.1.4
MarkupSafe         2.1.5
multidict          6.0.4
newsapi-python     0.2.7
openai             1.35.2
packaging          23.2
pip                24.0
pydantic           2.5.3
pydantic_core      2.14.6
python-dotenv      1.0.1
replicate          0.23.0
requests           2.31.0
sniffio            1.3.0
tqdm               4.66.1
typing_extensions  4.9.0
urllib3            2.1.0
Werkzeug           3.0.3
yarl               1.9.4